lwip: Use definitions from tools
This alters lwip to use network structs from the toolchain headers
instead of redefining them slightly differently itself. This is
important for being able to port code to work on lwip as most ported
code will depend on the tools headers instead of being written
specifically for lwIP and the mismatches caused result in memory
corruption and other problems. This also brings the changes in these
lwIP headers in line with the add-only code import philosophy.